Adds styling for locked names
