+++ src/login/styles/_Emails.scss
@@ -25,10 +25,10 @@
   margin-top: 2rem;
 }
 
-#personaldataview-form {
-  display: block;
-  width: 100%;
-}
+// #personaldataview-form {
+//   display: block;
+//   width: 100%;
+// }
 
 #suggested-password {
   display: table;
@@ -39,10 +39,10 @@
   color: $orange-highlight;
 }
 
-#personal-data-form {
-  display: flex;
-  flex-direction: column;
-}
+// #personal-data-form {
+//   display: flex;
+//   flex-direction: column;
+// }
 
 
 #add-more-button {
