+++ src/login/styles/_buttons.scss
@@ -7,6 +7,12 @@
   margin-top: 0;
   margin-bottom: 0;
 }
+
+.btn.disabled,
+.btn:disabled {
+  opacity: 1;
+}
+
 a.btn,
 button,
 button.btn {
@@ -70,6 +76,7 @@ button.btn-primary:hover {
   &:disabled,
   .disabled {
     background-color: $dark-gray;
+    // color: $light-gray;
     border: none;
     transform: scale(1);
   }
@@ -242,13 +249,26 @@ button.btn-link {
 }
 
 // retry/reload icon button
-button.icon {
-  height: auto;
-  svg {
-    color: $black;
-    height: 1rem;
-    width: 1rem;
-    transform: rotate(0);
+.icon-text {
+  button,
+  button:active,
+  button:focus,
+  button:active:focus {
+    border: transparent solid 2px;
+    margin-top: 0;
+    padding: 0.75rem;
+    background-color: white;
+    border-radius: 50%;
+    &.icon {
+      background-color: white;
+      height: auto;
+      svg {
+        color: $black;
+        height: 1rem;
+        width: 1rem;
+        transform: rotate(0);
+      }
+    }
   }
 }
