+++ src/login/styles/_inputs.scss
@@ -23,30 +23,38 @@ input[type="radio"] {
   align-self: center;
   /* background-color only for content */
   background-clip: content-box;
-  border: 1px solid $black;
-  background-color: $white;
+  box-shadow: 0 5px 17px rgba(89, 89, 89, 0.07);
+  border: 2px solid $orange-highlight;
+  // border: 1px solid $black;
+  //  background-color: $white;
+  background-color: transparent;
   border-radius: 50%;
 }
 
 /* appearance for checked radiobutton */
 input[type="radio"]:checked {
-  background-color: #28a745;
+  // background-color: #28a745;
+  background-color: $orange-highlight;
   border: none;
 }
 
 input[type="radio"]:checked.error {
-  background-color: $orange-highlight;
+  background: $dark-orange;
+  color: white;
+  opacity: 0.8;
+  border: none;
+  // background-color: $orange-highlight;
 }
 
 /* optional styles for centering radiobuttons */
 .radio-input-container {
   display: flex;
-  flex-wrap: wrap;
-  align-items: center;
   height: 3rem;
-  width: 100%;
-  margin-left: -2px;
-
+  // display: flex;
+  // flex: 1;
+  // align-items: center;
+  // height: 3rem;
+  // margin-left: -2px;
   span {
     font-size: 1.5rem;
     align-self: center;
