+++ src/login/styles/index.scss
@@ -29,5 +29,6 @@
 @import "Splash";
 @import "Errors";
 @import "Login";
+@import "PersonalData";
 @import "SuccessIcon";
 @import "../../../node_modules/bootstrap/dist/css/bootstrap.min.css";
