+++ src/login/styles/variables.scss
@@ -7,6 +7,7 @@ $black: #161616;
 $gray: #7c7c7c;
 $valid-green: #28a745;
 
+$dark-orange:#d3452e;
 $orange-highlight: #ff4500;
 $orange-opaque: #ffe7dd;
